fastapi_service/pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Lennox Stevenson <lennox@cobrand.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["all"], version = "^0.104.1" }
requests = "^2.31.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-api = "^1.21.0"
opentelemetry-instrumentation-aws-lambda = "^0.42b0"
opentelemetry-instrumentation-dbapi = "^0.42b0"
opentelemetry-instrumentation-logging = "^0.42b0"
opentelemetry-instrumentation-sqlite3 = "^0.42b0"
opentelemetry-instrumentation-urllib = "^0.42b0"
opentelemetry-instrumentation-wsgi = "^0.42b0"
opentelemetry-instrumentation-fastapi = "^0.42b0"
opentelemetry-instrumentation-httpx = "^0.42b0"
opentelemetry-instrumentation-jinja2 = "^0.42b0"
opentelemetry-instrumentation-tortoiseorm = "^0.42b0"
opentelemetry-instrumentation-asgi = "^0.42b0"
opentelemetry-exporter-otlp = "^1.21.0"
opentelemetry-instrumentation-grpc = "^0.42b0"
opentelemetry-instrumentation-requests = "^0.42b0"
opentelemetry-instrumentation-urllib3 = "^0.42b0"
opentelemetry-distro = { extras = ["otlp"], version = "^0.42b0" }
opentelemetry-exporter-gcp-trace = "^1.6.0"
opentelemetry-propagator-gcp = "^1.6.0"
opentelemetry-resourcedetector-gcp = "^1.6.0a0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
